requirements-docker.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-group base --no-group dev --format requirements-txt --output-file requirements-docker.txt --no-hashes
acres==0.5.0
    # via nipype
aiofiles==24.1.0
    # via
    #   frontend
    #   gradio
aistudio-sdk==0.3.8
    # via paddlex
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   gradio
    #   httpx
    #   starlette
asttokens==3.0.0
    # via stack-data
audioop-lts==0.2.2
    # via gradio
bce-python-sdk==0.9.52
    # via aistudio-sdk
brotli==1.2.0
    # via gradio
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
chardet==5.2.0
    # via paddlex
charset-normalizer==3.4.4
    # via requests
ci-info==0.3.0
    # via etelemetry
click==8.3.0
    # via
    #   aistudio-sdk
    #   nipype
    #   typer
    #   uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   colorlog
    #   ipython
    #   loguru
    #   tqdm
colorlog==6.10.1
    # via paddlex
configobj==5.0.9
    # via fitz
configparser==7.2.0
    # via fitz
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
decorator==5.2.1
    # via ipython
etelemetry==0.3.1
    # via nipype
executing==2.2.1
    # via stack-data
fastapi==0.128.0
    # via gradio
ffmpy==1.0.0
    # via gradio
filelock==3.20.0
    # via
    #   huggingface-hub
    #   modelscope
    #   nipype
    #   paddlex
    #   transformers
fitz==0.0.1.dev2
    # via ner-ocr
fonttools==4.60.1
    # via matplotlib
frontend==0.0.3
    # via ner-ocr
fsspec==2025.10.0
    # via
    #   gradio-client
    #   huggingface-hub
future==1.0.0
    # via bce-python-sdk
gradio==6.2.0
    # via ner-ocr
gradio-client==2.0.2
    # via gradio
groovy==0.1.2
    # via gradio
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.2.0 ; platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httplib2==0.31.0
    # via fitz
httpx==0.28.1
    # via
    #   gradio
    #   gradio-client
    #   langsmith
    #   safehttpx
huggingface-hub==0.36.0
    # via
    #   gradio
    #   gradio-client
    #   paddlex
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
imagesize==1.4.1
    # via paddlex
ipython==9.6.0
    # via ner-ocr
ipython-pygments-lexers==1.1.1
    # via ipython
itsdangerous==2.2.0
    # via frontend
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via gradio
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
kiwisolver==1.4.9
    # via matplotlib
langchain-core==1.0.5
    # via ner-ocr
langsmith==0.4.43
    # via langchain-core
levenshtein==0.27.3
    # via ner-ocr
loguru==0.7.3
    # via ner-ocr
looseversion==1.3.0
    # via nipype
lxml==6.0.2
    # via pyxnat
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   gradio
    #   jinja2
matplotlib==3.10.7
    # via ner-ocr
matplotlib-inline==0.2.1
    # via ipython
mdurl==0.1.2
    # via markdown-it-py
modelscope==1.32.0
    # via paddlex
networkx==3.5
    # via
    #   nipype
    #   prov
nibabel==5.3.2
    # via
    #   fitz
    #   nipype
nipype==1.10.0
    # via fitz
numpy==2.2.6
    # via
    #   contourpy
    #   fitz
    #   gradio
    #   matplotlib
    #   nibabel
    #   nipype
    #   opencv-contrib-python
    #   opencv-python
    #   paddlex
    #   pandas
    #   scipy
    #   shapely
    #   transformers
opencv-contrib-python==4.10.0.84
    # via paddlex
opencv-python==4.12.0.88
    # via ner-ocr
orjson==3.11.4
    # via
    #   gradio
    #   langsmith
packaging==25.0
    # via
    #   etelemetry
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   matplotlib
    #   nibabel
    #   nipype
    #   paddlex
    #   transformers
paddleocr==3.3.2
    # via ner-ocr
paddlex==3.3.9
    # via paddleocr
pandas==2.3.3
    # via
    #   fitz
    #   gradio
    #   paddlex
parso==0.8.5
    # via jedi
pathlib==1.0.1
    # via pyxnat
pdf2image==1.17.0
    # via ner-ocr
pexpect==4.9.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
    # via ipython
pillow==12.0.0
    # via
    #   gradio
    #   matplotlib
    #   paddlex
    #   pdf2image
prettytable==3.17.0
    # via
    #   aistudio-sdk
    #   paddlex
prompt-toolkit==3.0.52
    # via ipython
prov==2.1.1
    # via nipype
psutil==7.1.3
    # via aistudio-sdk
ptyprocess==0.7.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
    # via pexpect
pure-eval==0.2.3
    # via stack-data
puremagic==1.30
    # via nipype
py-cpuinfo==9.0.0
    # via paddlex
pyaml==25.7.0
    # via ner-ocr
pyclipper==1.3.0.post6
    # via paddlex
pycryptodome==3.23.0
    # via bce-python-sdk
pydantic==2.12.3
    # via
    #   fastapi
    #   gradio
    #   langchain-core
    #   langsmith
    #   ner-ocr
    #   paddlex
pydantic-core==2.41.4
    # via pydantic
pydot==4.0.1
    # via
    #   nipype
    #   prov
pydub==0.25.1
    # via gradio
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   rich
pyparsing==3.2.5
    # via
    #   httplib2
    #   matplotlib
    #   pydot
    #   rdflib
pypdfium2==5.0.0
    # via
    #   ner-ocr
    #   paddlex
python-bidi==0.6.7
    # via paddlex
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   nipype
    #   pandas
    #   prov
python-multipart==0.0.21
    # via gradio
pytz==2025.2
    # via pandas
pyxnat==1.6.3
    # via fitz
pyyaml==6.0.2
    # via
    #   gradio
    #   huggingface-hub
    #   langchain-core
    #   paddleocr
    #   paddlex
    #   pyaml
    #   transformers
rapidfuzz==3.14.3
    # via levenshtein
rdflib==7.4.0
    # via nipype
regex==2025.11.3
    # via transformers
requests==2.32.5
    # via
    #   aistudio-sdk
    #   etelemetry
    #   huggingface-hub
    #   langsmith
    #   modelscope
    #   paddleocr
    #   paddlex
    #   pyxnat
    #   requests-toolbelt
    #   transformers
requests-toolbelt==1.0.0
    # via langsmith
rich==14.2.0
    # via typer
ruamel-yaml==0.18.16
    # via paddlex
ruamel-yaml-clib==0.2.15 ; python_full_version < '3.14' and platform_python_implementation == 'CPython'
    # via ruamel-yaml
safehttpx==0.1.7
    # via gradio
safetensors==0.6.2
    # via transformers
scipy==1.16.3
    # via
    #   fitz
    #   nipype
semantic-version==2.10.0
    # via gradio
setuptools==80.9.0
    # via
    #   modelscope
    #   ner-ocr
shapely==2.1.2
    # via paddlex
shellingham==1.5.4
    # via typer
simplejson==3.20.2
    # via nipype
six==1.17.0
    # via
    #   bce-python-sdk
    #   python-dateutil
sniffio==1.3.1
    # via anyio
stack-data==0.6.3
    # via ipython
starlette==0.50.0
    # via
    #   fastapi
    #   frontend
    #   gradio
tenacity==9.1.2
    # via langchain-core
tokenizers==0.22.1
    # via transformers
tomlkit==0.13.3
    # via gradio
tqdm==4.67.1
    # via
    #   aistudio-sdk
    #   huggingface-hub
    #   modelscope
    #   transformers
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
traits==7.0.2
    # via nipype
transformers==4.57.1
    # via ner-ocr
typer==0.21.1
    # via gradio
types-pyyaml==6.0.12.20250915
    # via ner-ocr
typing-extensions==4.15.0
    # via
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   langchain-core
    #   paddleocr
    #   paddlex
    #   pydantic
    #   pydantic-core
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
ujson==5.11.0
    # via paddlex
urllib3==2.5.0
    # via
    #   modelscope
    #   requests
uvicorn==0.38.0
    # via
    #   frontend
    #   gradio
wcwidth==0.2.14
    # via
    #   prettytable
    #   prompt-toolkit
win32-setctime==1.2.0 ; sys_platform == 'win32'
    # via loguru
zstandard==0.25.0
    # via langsmith
